battle-log-hook-part-1

+++ platform/src/main/java/com/tarantula/game/GameRating.java
@@ -24,7 +24,7 @@ public class GameRating extends PlayerGameObject implements Rating {
         this.label = "rating";
     }
 
-    public Rating elo(boolean win,long opponentId,int teamPower){
+    public Rating elo(boolean win,long opponentId,long teamPower){
         //no direct call
         return this;
     }
